requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.7
django-bootstrap4==3.0.1
djangorestframework==3.12.4
Pillow==8.3.2
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
psycopg2==2.9.1
gunicorn
